Tweak .NET GoIO.Sensor_DDSMem_GetSensorNumber function prototype by changing sendQueryToHWFlag from byte to int to better match the GoIO_DLL C library function. Because of the way marshalling works, I suspect this change is a NOP. More hand waving.
[git-p4: depot-paths = "//depot/SDK/GoIO_SDK/": change = 35465]

+++ src/GoIOdotNET/GoIO.cs
@@ -673,7 +673,7 @@ namespace GoIOdotNET
 		public static extern Int32 Sensor_DDSMem_GetSensorNumber(
 			IntPtr hSensor,
 			out byte SensorNumber,
-			byte sendQueryToHardwareflag,
+			Int32 sendQueryToHardwareflag,
 			Int32 timeoutMs);
 
 		[DllImport("GoIO_DLL.dll", EntryPoint = "GoIO_Sensor_DDSMem_SetSerialNumber", CallingConvention = CallingConvention.Cdecl)]
